Add two modules for books manipulation
