+++ index.html
@@ -55,6 +55,6 @@
         <p>Lorem ipsum dolor sit amet consectetur adipisicing  nesciunt, doloremque adipisci delectus temporibus quia. Doloribus?</p>
     </footer>
 
-    <script src="script.js"></script>
+    <script type="module" src="script.js"></script>
 </body>
 </html>
